move DebugLogger to tasks/common
